refactor(commands): turn .run commands into tasks
