+++ packages/@angular/cli/commands/init.ts
@@ -26,7 +26,16 @@ const InitCommand: any = Command.extend({
   anonymousOptions: ['<glob-pattern>'],
 
   run: function (commandOptions: any, rawArgs: string[]) {
-    return require('./init.run').default.call(this, commandOptions, rawArgs);
+    const InitTask = require('../tasks/init').default;
+
+    const initTask = new InitTask({
+      cliProject: this.project,
+      project: this.project,
+      tasks: this.tasks,
+      ui: this.ui,
+    });
+
+    return initTask.run(commandOptions, rawArgs);
   }
 });
